New UX: Move pages to their own files
