+++ src/Components/App/CompanyChooserPage.tsx
@@ -1,0 +1,9 @@
+import { SetPageFunction } from './App';
+
+type CompanyChooserPageProps = {
+    setPage: SetPageFunction;
+};
+
+export const CompanyChooserPage = (props: CompanyChooserPageProps) => {
+    return <></>;
+};
